LayoutTests/accessibility/mac/select-text/select-text-135575-expected.txt

The quick brown fox jumps over the lazy dog.

TEXT2: THE quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog.

This tests the ability to select and replace text with respect to selection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS windowSelection.getRangeAt(0).toString() is 'The'
PASS result is 'THE'
PASS windowSelection.getRangeAt(0).toString() is 'THE'
PASS windowSelection.getRangeAt(0).startOffset is 7
PASS windowSelection.getRangeAt(0).endOffset is 10
PASS successfullyParsed is true

TEST COMPLETE
